web: move models into api folder
