+++ apps/web/src/lib/api/api.ts
@@ -1,5 +1,5 @@
 import { API_HOST, API_KEY } from "$env/static/private";
-import type { AppData } from "$lib/models/AppData";
+import type { AppData } from "$lib/api/models/AppData";
 
 export const fetchThings = async (fetch): Promise<AppData> => {
   const result = await fetch(`${API_HOST}/web/things`, {
