+++ apps/web/src/lib/api/models/ItemDetails.ts
