+++ apps/web/src/lib/api/models/ThingDetails.ts
