+++ apps/web/src/routes/(app)/+page.server.ts
@@ -1,5 +1,5 @@
-import type { AppData } from '$lib/models/AppData';
-import { fetchThings } from '$lib/server/api.js'
+import type { AppData } from '$lib/api/models/AppData';
+import { fetchThings } from '$lib/api/api.js'
 
 export const load = async ({ fetch }): Promise<AppData> => {
   return fetchThings(fetch);
